sumo_microservice/requirements.txt
traci==1.20.0
pymongo==4.6.0
redis==5.0.1
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.4
